- [ZBX-3747] Fixed compiler warnings and some style improvements
git-svn-id: svn://svn.zabbix.com/branches/dev@19345 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ ZBX-3747/include/common.h
@@ -837,7 +837,7 @@ int	comms_parse_response(char *xml, char *host, int host_len, char *key, int key
 		char *lastlogsize, int lastlogsize_len, char *timestamp, int timestamp_len,
 		char *source, int source_len, char *severity, int severity_len);
 
-int 	parse_command(const char *command, char *cmd, size_t cmd_max_len, char *param, int param_max_len);
+int 	parse_command(const char *command, char *cmd, size_t cmd_max_len, char *param, size_t param_max_len);
 
 typedef struct zbx_regexp_s
 {
